use code model in searchagent's code agent calls
