+++ app/src/test/java/ai/brokk/testutil/TestContextManager.java
@@ -10,7 +10,6 @@ import ai.brokk.analyzer.Languages;
 import ai.brokk.analyzer.ProjectFile;
 import ai.brokk.context.Context;
 import ai.brokk.git.TestRepo;
-import ai.brokk.prompts.EditBlockParser;
 import dev.langchain4j.model.chat.StreamingChatModel;
 import java.io.File;
 import java.nio.file.Path;
@@ -153,8 +152,9 @@ public final class TestContextManager implements IContextManager {
         return stubService;
     }
 
-    public EditBlockParser getParserForWorkspace() {
-        return EditBlockParser.instance;
+    @Override
+    public StreamingChatModel getCodeModel() {
+        return null;
     }
 
     /**
